TF-1810 Move all hive clients to a separate folder
(cherry picked from commit c72e16b576835b7748e20ab3e5cce5e1f505a65f)
