+++ lib/features/caching/clients/email_cache_client.dart
