+++ lib/features/caching/clients/hive_cache_version_client.dart
@@ -1,5 +1,5 @@
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:tmail_ui_user/features/caching/cache_version_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/cache_version_client.dart';
 import 'package:tmail_ui_user/main/exceptions/exception_thrower.dart';
 
 class HiveCacheVersionClient extends CacheVersionClient {
