+++ lib/features/caching/clients/recent_search_cache_client.dart
