+++ lib/features/mailbox/data/local/mailbox_cache_manager.dart
@@ -4,7 +4,7 @@ import 'package:jmap_dart_client/jmap/core/user_name.dart';
 import 'package:jmap_dart_client/jmap/mail/mailbox/mailbox.dart';
 import 'package:model/extensions/account_id_extensions.dart';
 import 'package:model/mailbox/presentation_mailbox.dart';
-import 'package:tmail_ui_user/features/caching/mailbox_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/mailbox_cache_client.dart';
 import 'package:tmail_ui_user/features/mailbox/data/extensions/list_mailbox_cache_extension.dart';
 import 'package:tmail_ui_user/features/mailbox/data/extensions/list_mailbox_extension.dart';
 import 'package:tmail_ui_user/features/mailbox/data/extensions/list_mailbox_id_extension.dart';
