+++ lib/features/mailbox_dashboard/data/datasource_impl/search_datasource_impl.dart
@@ -1,5 +1,5 @@
 
-import 'package:tmail_ui_user/features/caching/recent_search_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/recent_search_cache_client.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/data/datasource/search_datasource.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/data/model/recent_search_cache.dart';
 import 'package:tmail_ui_user/features/mailbox_dashboard/domain/extensions/list_recent_search_extension.dart';
