+++ lib/features/push_notification/data/local/fcm_cache_manager.dart
@@ -2,8 +2,8 @@ import 'package:fcm/model/type_name.dart';
 import 'package:jmap_dart_client/jmap/account_id.dart';
 import 'package:jmap_dart_client/jmap/core/user_name.dart';
 import 'package:model/extensions/account_id_extensions.dart';
-import 'package:tmail_ui_user/features/caching/fcm_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/subscription_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/fcm_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/subscription_cache_client.dart';
 import 'package:tmail_ui_user/features/caching/utils/cache_utils.dart';
 import 'package:tmail_ui_user/features/push_notification/data/model/fcm_subscription.dart';
 import 'package:tmail_ui_user/features/push_notification/domain/exceptions/fcm_exception.dart';
