+++ lib/main/bindings/local/local_bindings.dart
@@ -1,20 +1,20 @@
 
 import 'package:get/get.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:tmail_ui_user/features/caching/account_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/authentication_info_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/account_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/authentication_info_cache_client.dart';
 import 'package:tmail_ui_user/features/caching/caching_manager.dart';
-import 'package:tmail_ui_user/features/caching/email_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/encryption_key_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/fcm_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/hive_cache_version_client.dart';
-import 'package:tmail_ui_user/features/caching/mailbox_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/recent_login_url_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/recent_login_username_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/recent_search_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/state_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/subscription_cache_client.dart';
-import 'package:tmail_ui_user/features/caching/token_oidc_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/email_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/encryption_key_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/fcm_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/hive_cache_version_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/mailbox_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/recent_login_url_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/recent_login_username_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/recent_search_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/state_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/subscription_cache_client.dart';
+import 'package:tmail_ui_user/features/caching/clients/token_oidc_cache_client.dart';
 import 'package:tmail_ui_user/features/cleanup/data/local/recent_login_url_cache_manager.dart';
 import 'package:tmail_ui_user/features/cleanup/data/local/recent_login_username_cache_manager.dart';
 import 'package:tmail_ui_user/features/cleanup/data/local/recent_search_cache_manager.dart';
